move utils tests to seperate folder
